+++ projects/sol-wallet/go.mod
@@ -3,7 +3,7 @@ module solana-wallet
 go 1.13
 
 require (
-	github.com/portto/solana-go-sdk v1.16.1 // indirect
+	github.com/portto/solana-go-sdk v1.18.0 // indirect
 	github.com/stretchr/testify v1.7.1 // indirect
 	github.com/tyler-smith/go-bip32 v1.0.0 // indirect
 	github.com/tyler-smith/go-bip39 v1.1.0 // indirect
